# openapi.json
{
  "components": {
    "schemas": {
...
      "NodeResponse": {
        "properties": {
          "created_at": {
            "description": "RFC 3339 UTC.",
            "type": "string"
          },
          "display_name": {
            "description": "Mutable, human-readable name.",
            "type": "string"
          },
          "fingerprint": {
            "description": "Cert fingerprint (cluster-CA-signed).",
            "type": "string"
          },
          "id": {
            "description": "Stable node UUID.",
            "type": "string"
          },
          "last_seen": {
            "description": "RFC 3339 UTC, last activity stamp. `None` on a fresh registration.",
            "nullable": true,
            "type": "string"
          },
          "online": {
            "description": "`true` when `status == \"active\"`.",
            "type": "boolean"
          },
          "role": {
            "description": "`\"node\"` or `\"admin\"` (and the implied `\"control\"` superset).",
            "type": "string"
          },
          "status": {
            "description": "`\"active\"` or `\"revoked\"`.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "role",
          "status",
          "online",
          "fingerprint",
          "created_at"
        ],
        "type": "object"
      },
      "SetNameRequest": {
        "properties": {
          "display_name": {
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "type": "object"
      },
      "SetRoleRequest": {
        "properties": {
          "role": {
            "description": "`\"node\"` or `\"admin\"`.",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API for managing the MLSH overlay network.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "MLSH Control API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/nodes": {
      "get": {
        "operationId": "list_nodes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NodeResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Nodes"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "List every node in the cluster.",
        "tags": [
          "Nodes"
        ]
      }
    },
    "/api/v1/nodes/{node}": {
      "delete": {
        "operationId": "delete_node",
        "parameters": [
          {
            "description": "Node UUID or display name",
            "in": "path",
            "name": "node",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Node revoked"
          },
          "404": {
            "description": "Node not found"
          }
        },
        "summary": "Delete (= revoke) a node. Idempotent.",
        "tags": [
          "Nodes"
        ]
      },
      "get": {
        "operationId": "get_node",
        "parameters": [
          {
            "description": "Node UUID or display name",
            "in": "path",
            "name": "node",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeResponse"
                }
              }
            },
            "description": "Node"
          },
          "404": {
            "description": "Node not found"
          }
        },
        "summary": "Get a single node by id (or display name).",
        "tags": [
          "Nodes"
        ]
      }
    },
    "/api/v1/nodes/{node}/name": {
      "post": {
        "operationId": "set_name",
        "parameters": [
          {
            "description": "Node UUID or display name",
            "in": "path",
            "name": "node",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Name updated"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Node not found"
          }
        },
        "summary": "Set the node's display name.",
        "tags": [
          "Nodes"
        ]
      }
    },
    "/api/v1/nodes/{node}/revoke": {
      "post": {
        "operationId": "revoke",
        "parameters": [
          {
            "description": "Node UUID or display name",
            "in": "path",
            "name": "node",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Node revoked"
          },
          "404": {
            "description": "Node not found"
          }
        },
        "summary": "Revoke a node. Idempotent — succeeds even if the node was already revoked.",
        "tags": [
          "Nodes"
        ]
      }
    },
    "/api/v1/nodes/{node}/role": {
      "post": {
        "operationId": "set_role",
        "parameters": [
          {
            "description": "Node UUID or display name",
            "in": "path",
            "name": "node",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Role updated"
          },
          "400": {
            "description": "Invalid role"
          },
          "404": {
            "description": "Node not found"
          }
        },
        "summary": "Set the node's role (`node` or `admin`).",
        "tags": [
          "Nodes"
        ]
      }
    },
...
  }
}